+++ noodles-bam/CHANGELOG.md
@@ -4,9 +4,9 @@
 
 ### Added
 
-  * bam/lazy/record: Add mapping quality (`lazy::record::MappingQaulity`),
-    position (`lazy::record::Position`) and reference sequence ID
-    (`lazy::record::ReferenceSequenceId`) wrappers.
+  * bam/lazy/record: Add flags (`lazy::record::Flags`), mapping quality
+    (`lazy::record::MappingQaulity`), position (`lazy::record::Position`) and
+    reference sequence ID (`lazy::record::ReferenceSequenceId`) wrappers.
 
   * bam/lazy/record/data/field/value/array: Add values wrapper (`Values`).